Add Village Name + Link to Village to Feed and Inbox + Denormalise Village properties in Post
Added village name and slug to Post schema and added the village name
and slug when a post gets inserted (via a hook). Next to that I redid
and earlier commit “Fix that only posts relevant to Village are on
Village View” because now we have the villageSlug in post and can do
this with less resources.